Fix int signage comparison warning in AudioBuffer.

+++ include/bricks/audio/audiobuffer.h
@@ -59,7 +59,7 @@ namespace Bricks { namespace Audio {
 		int CopyTo(AudioBuffer<AudioSample>& dest, u32 sourcechannel = 0, u32 sourceindex = 0, u32 destchannel = 0, u32 destindex = 0, s32 count = -1) const {
 			if (count < 0)
 				count = size;
-			count = BRICKS_FEATURE_MIN(BRICKS_FEATURE_MIN(count, size - sourceindex), dest.GetSize() - destindex);
+			count = BRICKS_FEATURE_MIN(BRICKS_FEATURE_MIN((u32)count, size - sourceindex), dest.GetSize() - destindex);
 			for (u32 i = sourcechannel; i < channels; i++)
 				memcpy(dest[destchannel + i - sourcechannel] + destindex, buffer[i] + sourceindex, count * sizeof(AudioSample));
 			return count;
